src/suppliers/mouser.py
import os
from typing import Dict, Optional, Tuple
import requests
from bs4 import BeautifulSoup
from .base import SupplierBase
from dotenv import load_dotenv

load_dotenv()

class MouserSupplier(SupplierBase):
    def __init__(self):
        self.api_key = os.getenv('MOUSER_API_KEY')
        self.base_url = "https://api.mouser.com/api/v1"
        
    def get_part_info(self, part_number: str) -> Optional[Tuple[float, int]]:
        """
        Get price and stock information from Mouser.
        Requires API key setup in .env file.
        """
        # Note: Implementation needs Mouser API credentials
        # This is a placeholder for the actual API implementation
        return None